Fixed order of itineraries by adding order

+++ Neo4j-Example/src/main/java/com/example/Neo4jExample/dto/PoiDTO.java
@@ -14,7 +14,7 @@ public class PoiDTO {
     private String description;
     private CoordinateDTO coordinate;
     private TimeSlotDTO hours;
-    private Integer timeToVisit;
+    private Double timeToVisit;
     private AddressDTO address;
     private Double ticketPrice;
     private Collection<String> contributors;
